ocr-images-novel/requirements.txt
pytesseract
pillow
easyocr
paddlepaddle
paddleocr
